Added vertex and fragment shaders for entities.
They will be used to render game::entity::Entity objects using OpenGL instancing.
In the vertex shader a little trick is used to make sure in the fragment shader the relative texture position is rounded to the right pixel. Without this the textures would sometimes be off by one pixel.