Fixed wrong rendering sizes.
The movables.vert shader was adding the entire size of every object to every side, doubling the real size. game::ent::Item's, which sizes are hardcoded into the data passed to the shader, had a wrong size value inside sizeX() and sizeY(). Also doubled game::ent::mov::Rocket size to match the one that appared before the shader was fixed.