Completely remade misc::Acceleration.
Before the concept worked only with one acceleration, because the Acceleration class accumulated the velocity resulting in wrong behaviour when changing accelerations.
Now only Δv (velocity difference) can be calculated, based on Δt (time difference): the velocity from the beginning of the acceleration can't. The user can then sum the velocity of every acceleration and multiply the result again by Δt to get Δs (space difference). Acceleration now also keeps track of the rotation and can be decomposed into x and y.
To make it simpler to sum accelerations there is now AccelerationSum, that results from +/- operations on Acceleration's.