Closes #1
Remade keys input handling. Now there is an array containing the status of all keys (both from keyboard and mouse). It is updated every frame by checking every key. This makes sure that if a press/release is shorter than a frame (thus the return value of glfwGetKey() is PRESS only the first time it is called) the status of the key is the same for all listeners.